requirements.txt
requests
python-dotenv
openai
sqlalchemy
psycopg2-binary
torch
numpy
pandas
pyarrow
scikit-learn
